+++ HaselTweaks/Enums/PluginLanguageOverride.cs
@@ -1,0 +1,8 @@
+namespace HaselTweaks.Enums;
+
+public enum PluginLanguageOverride
+{
+    None,
+    Dalamud,
+    Client
+}