+++ HaselTweaks/Utils/ImGui/IconButton.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Numerics;
 using Dalamud.Interface;
 using Dalamud.Interface.Raii;
@@ -13,20 +14,29 @@ public partial class ImGuiUtils
         return ImGui.CalcTextSize(icon.ToIconString()) + ImGui.GetStyle().FramePadding * 2;
     }
 
-    public static bool IconButton(string key, FontAwesomeIcon icon, string tooltip, Vector2 size = default, bool disabled = false)
+    public static bool IconButton(string key, FontAwesomeIcon icon, string tooltip, Vector2 size = default, bool disabled = false, bool active = false)
     {
         using var iconFont = ImRaii.PushFont(UiBuilder.IconFont);
         if (!key.StartsWith("##")) key = "##" + key;
 
-        var a = disabled ? ImRaii.PushColor(ImGuiCol.Text, ImGui.GetStyle().Colors[(int)ImGuiCol.TextDisabled]) : null;
-        var b = disabled ? ImRaii.PushColor(ImGuiCol.ButtonActive, ImGui.GetStyle().Colors[(int)ImGuiCol.Button]) : null;
-        var c = disabled ? ImRaii.PushColor(ImGuiCol.ButtonHovered, ImGui.GetStyle().Colors[(int)ImGuiCol.Button]) : null;
+        var disposables = new List<IDisposable>();
+
+        if (disabled)
+        {
+            disposables.Add(ImRaii.PushColor(ImGuiCol.Text, ImGui.GetStyle().Colors[(int)ImGuiCol.TextDisabled]));
+            disposables.Add(ImRaii.PushColor(ImGuiCol.ButtonActive, ImGui.GetStyle().Colors[(int)ImGuiCol.Button]));
+            disposables.Add(ImRaii.PushColor(ImGuiCol.ButtonHovered, ImGui.GetStyle().Colors[(int)ImGuiCol.Button]));
+        }
+        else if (active)
+        {
+            disposables.Add(ImRaii.PushColor(ImGuiCol.Button, ImGui.GetStyle().Colors[(int)ImGuiCol.ButtonActive]));
+        }
 
         var pressed = ImGui.Button(icon.ToIconString() + key, size);
 
-        c?.Dispose();
-        b?.Dispose();
-        a?.Dispose();
+        foreach (var disposable in disposables)
+            disposable.Dispose();
+
         iconFont?.Dispose();
 
         if (ImGui.IsItemHovered())